[RISCV][LLD] Zcmt RISC-V extension in lld
This patch implements optimizations for the zcmt extension in lld. A new
TableJumpSection has been added. Scans each R_RISCV_CALL/R_RISCV_CALL_PLT
relocType in each section before the linker relaxation, recording the symbol

In finalizeContents the recorded symbol names are sorted in descending order by
the number of jumps. The top symbols are compressed to table jumps during the
relax process.

This is a continuation of PR #77884

Co-authored-by: Craig Topper <craig.topper@sifive.com>
Co-authored-by: VincentWu <43398706+Xinlong-Wu@users.noreply.github.com>
Co-authored-by: Scott Egerton <9487234+ScottEgerton@users.noreply.github.com>